pyproject.toml
[project]
name = "automatereportsystem"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosmtplib>=4.0.1",
    "aiosqlite>=0.21.0",
    "alembic>=1.16.4",
    "apscheduler>=3.11.0",
    "asyncio>=3.4.3",
    "bcrypt==3.2.0",
    "cryptography>=45.0.5",
    "fastapi[all]>=0.116.1",
    "pandas>=2.0.0",
    "passlib[bcrypt]>=1.7.4",
    "plotly>=5.0.0",
    "psutil>=7.0.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.41",
    "streamlit>=1.47.0",
]
